requirements.txt
flask
flask-sqlalchemy
pypdf
groq
python-dotenv
gunicorn
